src/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==3.2
django-froala-editor==4.0.18
docker==6.0.1
idna==3.4
packaging==23.1
pytz==2023.2
pywin32==306
requests==2.29.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
websocket-client==1.5.1
coverage==7.2.3
docker==6.0.1